Add `pip install -U` to installation instructions
The instructions for running the Vagrant machine as written lead to an outdated dependency being installed, which leaves the local CourtListener instance broken. I'm not sure how to do it the "right" way with Ansible, so I just added an instruction to the README file to upgrade the dependencies of the program.